Tests for executable
