+++ supernova/config.py
@@ -36,24 +36,6 @@ def run_config(config_file_override=None):
     return nova_creds
 
 
-def check_environment_presets():
-    """
-    Checks for environment variables that can cause problems with supernova
-    """
-    presets = [x for x in os.environ.copy().keys() if x.startswith('NOVA_') or
-               x.startswith('OS_')]
-    if len(presets) < 1:
-        return True
-    else:
-        print("_" * 80)
-        print("*WARNING* Found existing environment variables that may "
-              "cause conflicts:")
-        for preset in presets:
-            print("  - %s" % preset)
-        print("_" * 80)
-        return False
-
-
 def load_config(config_file_override=None):
     """
     Pulls the supernova configuration file and reads it
